requirements.txt
django
djangorestframework
python-slugify
Pillow
drf-yasg
django-cors-headers
python-decouple
django-filter

celery
redis

secure-smtplib
